+++ src/MudBlazor/Services/Scroll/IScrollListenerFactory.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace MudBlazor;
+
+public interface IScrollListenerFactory
+{
+    IScrollListener Create(string selector);
+}
