--- src/MudBlazor/Services/Scroll/ScrollManagerException.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Diagnostics.CodeAnalysis;
-
-namespace MudBlazor
-{
-
-    [ExcludeFromCodeCoverage]
-    public class ScrollManagerException : Exception
-    {
-        public ScrollManagerException() { }
-        public ScrollManagerException(string message) : base(message) { }
-        public ScrollManagerException(string message, Exception inner) : base(message, inner) { }
-    }
-}
