+++ src/MudBlazor/Services/Scroll/ScrollSectionCenteredEventArgs.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace MudBlazor;
+
+public class ScrollSectionCenteredEventArgs
+{
+    public string Id { get; }
+
+    public ScrollSectionCenteredEventArgs(string id)
+    {
+        Id = id;
+    }
+}
