+++ src/MudBlazor/Services/Scroll/ScrollSpyFactory.cs
@@ -2,17 +2,11 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.JSInterop;
 
 namespace MudBlazor
 {
-    public interface IScrollSpyFactory
-    {
-        IScrollSpy Create();
-    }
-
     public class ScrollSpyFactory : IScrollSpyFactory
     {
         private readonly IServiceProvider _provider;
